Refactor branch

+++ TAABP.Application/Services/ReservationService.cs
@@ -2,6 +2,8 @@
 using TAABP.Application.Exceptions;
 using TAABP.Application.Profile.ReservationMapping;
 using TAABP.Application.RepositoryInterfaces;
+using TAABP.Application.ServiceInterfaces;
+using TAABP.Core;
 
 namespace TAABP.Application.Services
 {
@@ -11,19 +13,28 @@ namespace TAABP.Application.Services
         private readonly IReservationMapper _reservationMapper;
         private readonly IRoomRepository _roomRepository;
         private readonly IUserRepository _userRepository;
+        private readonly IHotelRepository _hotelRepository;
+        private readonly ICityRepository _cityRepository;
 
         public ReservationService(IReservationRepository reservationRepository, IReservationMapper reservationMapper, 
-            IRoomRepository roomRepository, IUserRepository userRepository)
+            IRoomRepository roomRepository, IUserRepository userRepository, IHotelRepository hotelRepository,
+            ICityRepository cityRepository)
         {
             _reservationRepository = reservationRepository;
             _reservationMapper = reservationMapper;
             _roomRepository = roomRepository;
             _userRepository = userRepository;
+            _hotelRepository = hotelRepository;
+            _cityRepository = cityRepository;
         }
 
         public async Task<ReservationDto> GetReservationByIdAsync(int id)
         {
             var reservation = await _reservationRepository.GetReservationByIdAsync(id);
+            if (reservation == null)
+            {
+                throw new EntityNotFoundException("Reservation not found");
+            }
             return _reservationMapper.ReservationToResevationDto(reservation);
         }
 
@@ -33,7 +44,7 @@ namespace TAABP.Application.Services
             return reservations.Select(r => _reservationMapper.ReservationToResevationDto(r)).ToList();
         }
 
-        public async Task CreateReservationAsync(ReservationDto reservationDto)
+        public async Task<int> CreateReservationAsync(ReservationDto reservationDto)
         {
             var room = await _roomRepository.GetRoomByIdAsync(reservationDto.RoomId);
             var user = await _userRepository.GetUserByIdAsync(reservationDto.UserId);
@@ -41,10 +52,31 @@ namespace TAABP.Application.Services
             {
                 throw new EntityNotFoundException("Room or user not found");
             }
+            if (!room.IsAvailable)
+            {
+                throw new InvalidOperationException("Room is not available");
+            }
             var reservation = _reservationMapper.ReservationDtoToReservation(reservationDto);
+            if (reservation.StartDate < DateTime.Now)
+            {
+                throw new InvalidOperationException("Start date cannot be in the past");
+            }
+            if (reservation.StartDate > reservation.EndDate)
+            {
+                throw new InvalidOperationException("Start date must be before end date");
+            }
+            var hotel = await _hotelRepository.GetHotelByIdAsync(room.HotelId);
+            await _roomRepository.BookRoomAsync(room.RoomId);
+            await _hotelRepository.IncrementNumberOfVisitsAsync(room.HotelId);
+            await _cityRepository.IncrementNumberOfVisitsAsync(hotel.CityId);
             var price = (reservation.EndDate - reservation.StartDate).Days * room.PricePerNight;
+            if(price == 0)
+            {
+                price = room.PricePerNight;
+            }
             reservation.Price = price;
             await _reservationRepository.CreateReservationAsync(reservation);
+            return reservation.ReservationId;
         }
 
         public async Task UpdateReservationAsync(ReservationDto reservationDto)
@@ -54,7 +86,26 @@ namespace TAABP.Application.Services
             {
                 throw new EntityNotFoundException("Reservation not found");
             }
+            if (DateTime.Now >= targetReservation.StartDate.AddHours(-24) || DateTime.Now >= targetReservation.EndDate)
+            {
+                throw new InvalidOperationException("Reservations cannot be updated within 24 hours of the start date, during the stay, or after completion.");
+            }
+            if (reservationDto.StartDate < DateTime.Now)
+            {
+                throw new InvalidOperationException("Start date cannot be in the past");
+            }
             var reservation = _reservationMapper.ReservationDtoToReservation(reservationDto);
+            var room = await _roomRepository.GetRoomByIdAsync(reservation.RoomId);
+            if(room == null)
+            {
+                throw new EntityNotFoundException("Room not found");
+            }
+            var price = (reservation.EndDate - reservation.StartDate).Days * room.PricePerNight;
+            if (price == 0)
+            {
+                price = room.PricePerNight;
+            }
+            reservation.Price = price;
             await _reservationRepository.UpdateReservationAsync(reservation);
         }
 
@@ -65,6 +116,15 @@ namespace TAABP.Application.Services
             {
                 throw new EntityNotFoundException("Reservation not found");
             }
+            if (DateTime.Now >= targetReservation.StartDate.AddHours(-24) || DateTime.Now >= targetReservation.EndDate)
+            {
+                throw new InvalidOperationException("Reservations cannot be canceled within 24 hours of the start date, during the stay, or after completion.");
+            }
+            var room = await _roomRepository.GetRoomByIdAsync(targetReservation.RoomId);
+            await _roomRepository.UnbookRoomAsync(room.RoomId);
+            var hotel = await _hotelRepository.GetHotelByIdAsync(room.HotelId);
+            await _hotelRepository.DecrementNumberOfVisitsAsync(hotel.HotelId);
+            await _cityRepository.DecrementNumberOfVisitsAsync(hotel.CityId);
             await _reservationRepository.DeleteReservationAsync(targetReservation);
         }
     }
